Centralize color conversion functions in Common/ColorConv.
