+++ Globals.h
@@ -30,24 +30,6 @@
 #define IS_LITTLE_ENDIAN (*(const u16 *)"\0\xff" >= 0x100)
 #define IS_BIG_ENDIAN (*(const u16 *)"\0\xff" < 0x100)
 
-inline u8 Convert4To8(u8 v)
-{
-	// Swizzle bits: 00001234 -> 12341234
-	return (v << 4) | (v);
-}
-
-inline u8 Convert5To8(u8 v)
-{
-	// Swizzle bits: 00012345 -> 12345123
-	return (v << 3) | (v >> 2);
-}
-
-inline u8 Convert6To8(u8 v)
-{
-	// Swizzle bits: 00123456 -> 12345612
-	return (v << 2) | (v >> 4);
-}
-
 static inline u8 clamp_u8(int i) {
 #ifdef ARM
 	asm("usat %0, #8, %1" : "=r"(i) : "r"(i));
