+++ headless/Compare.cpp
@@ -18,6 +18,7 @@
 #include "headless/Compare.h"
 #include "file/file_util.h"
 #include "Core/Host.h"
+#include "Common/ColorConv.h"
 #include "GPU/Common/GPUDebugInterface.h"
 #include "GPU/Common/TextureDecoder.h"
